+++ TinkoffASDKCore/TinkoffASDKCore/Domain/ExternalAPI/Payloads/GetSBPBanksPayload.swift
@@ -38,4 +38,8 @@ public struct GetSBPBanksPayload: Decodable {
         }
         banks = resultBanks
     }
+
+    public init(banks: [SBPBank]) {
+        self.banks = banks
+    }
 }